refactor: Move conversion module to its own file.
